requirements.txt
# ── Core ──────────────────────────────────────────────────────────────────────
streamlit>=1.35.0

# ── LLM Providers ─────────────────────────────────────────────────────────────
openai>=1.30.0
groq>=0.9.0
google-generativeai>=0.7.0

# ── RAG / Embeddings ──────────────────────────────────────────────────────────
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
numpy>=1.26.0

# ── Document Processing ───────────────────────────────────────────────────────
pdfplumber>=0.11.0
python-docx>=1.1.0
pypdf>=4.2.0

# ── Web Search ────────────────────────────────────────────────────────────────
requests>=2.32.0

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
